petクラスの作成
